+++ studio/src/app/pages/editor/app-editor/app-editor.tsx
@@ -75,6 +75,12 @@ export class AppEditor {
     @State()
     private hideNavigation: boolean = false;
 
+    @State()
+    private slidesEditable: boolean = false;
+
+    @State()
+    private fullscreen: boolean = false;
+
     constructor() {
         this.authService = AuthService.getInstance();
         this.anonymousService = AnonymousService.getInstance();
@@ -126,8 +132,14 @@ export class AppEditor {
             // Hide actions footer till deck is editable
             this.hideFooterActions = false;
 
+            this.slidesEditable = true;
+
             await this.contentEditable(slide);
+
+            await this.showHelp();
         });
+
+        this.fullscreen = DeckDeckGoUtils.isFullscreen();
     }
 
     async destroy() {
@@ -459,6 +471,10 @@ export class AppEditor {
 
             await (deck as any).toggleFullScreen();
 
+            setTimeout(() => {
+                this.fullscreen = DeckDeckGoUtils.isFullscreen();
+            }, 100);
+
             resolve();
         });
     }
@@ -512,6 +528,17 @@ export class AppEditor {
         });
     }
 
+    private showHelp(): Promise<void> {
+        return new Promise<void>(async (resolve) => {
+           const help: HTMLElement = this.el.querySelector('app-help');
+           if (help) {
+               await (help as any).displayHelp();
+           }
+
+           resolve();
+        });
+    }
+
     private stickyToolbarActivated($event: CustomEvent) {
         this.hideFooterActions = $event ? $event.detail : false;
         this.hideNavigation = $event ? DeckDeckGoUtils.isIOS() && $event.detail : false;
@@ -521,8 +548,7 @@ export class AppEditor {
         return [
             <app-navigation publish={true} class={this.hideNavigation ? 'hidden' : undefined}></app-navigation>,
             <ion-content class="ion-padding">
-                <app-help class={!this.slidesFetched ? 'hidden' : undefined}></app-help>
-                <main class={this.slidesFetched ? (this.presenting ? 'ready idle' : 'ready') : undefined}>
+                <main class={this.slidesEditable ? (this.presenting ? 'ready idle' : 'ready') : undefined}>
 
                     {this.renderLoading()}
 
@@ -544,7 +570,8 @@ export class AppEditor {
                 <ion-toolbar>
                     <ion-buttons slot="start" class={this.hideFooterActions ? 'hidden' : undefined}>
                         <app-add-slide-action
-                            onActionOpenSlideAdd={($event: CustomEvent) => this.onActionOpenSlideAdd($event)}></app-add-slide-action>
+                            onActionOpenSlideAdd={($event: CustomEvent) => this.onActionOpenSlideAdd($event)}>
+                        </app-add-slide-action>
 
                         <ion-tab-button onClick={() => this.animatePrevNextSlide(false)} color="primary" mode="md">
                             <ion-icon name="arrow-back"></ion-icon>
@@ -562,8 +589,7 @@ export class AppEditor {
                         </ion-tab-button>
 
                         <ion-tab-button onClick={() => this.toggleFullScreen()} color="primary" class="wider-devices" mode="md">
-                            <ion-icon name="expand"></ion-icon>
-                            <ion-label>Fullscreen</ion-label>
+                            {this.renderFullscreen()}
                         </ion-tab-button>
 
                         <ion-tab-button onClick={() => this.openRemoteControl()} color="primary" class="wider-devices" mode="md">
@@ -575,6 +601,10 @@ export class AppEditor {
                             <ion-icon name="more" md="md-more" ios="md-more"></ion-icon>
                             <ion-label>More</ion-label>
                         </ion-tab-button>
+                    </ion-buttons>
+
+                    <ion-buttons slot="end" class={this.hideFooterActions ? 'hidden' : undefined}>
+                        <app-help></app-help>
                     </ion-buttons>
                 </ion-toolbar>
             </ion-footer>,
@@ -591,6 +621,20 @@ export class AppEditor {
             return undefined;
         } else {
             return <ion-spinner color="primary"></ion-spinner>;
+        }
+    }
+
+    private renderFullscreen() {
+        if (this.fullscreen) {
+            return [
+                <ion-icon name="contract"></ion-icon>,
+                <ion-label>Windowed</ion-label>
+            ];
+        } else {
+            return [
+                <ion-icon name="expand"></ion-icon>,
+                <ion-label>Fullscreen</ion-label>
+            ];
         }
     }
 }